requirements.txt
cffi>=0.9.2
Flask>=0.10.1
Flask-Bootstrap>=3.3.2.1
pycparser>=2.10
pyalsaaudio>=0.8
gevent>=1.0.1
